requirements.txt
streamlit
google-generative-ai
nylas
pinecone-client
pandas
matplotlib
streamlit-extras
python-dotenv
